fix: correct integer division in SRT session summary stats
byteRetransTotal, byteSndDropTotal and byteSentTotal are all uint64_t,
so the ratio was computed with integer division before being scaled by
100.0. Any value below 1.0 truncated to 0, making the summary print
0.0% for every run with less than 100% retransmission.

Cast to double before dividing. The byteSentTotal == 0 guard is
unchanged.

+++ plugins/obs-ffmpeg/obs-ffmpeg-srt.h
@@ -902,8 +902,8 @@ static int libsrt_close(URLContext *h)
 	     "\tbytes retransmitted [%.1f %%]\n"
 	     "\tbytes dropped [%.1f %%]\n",
 	     (double)perf.msTimeStamp / 1000.0, perf.mbpsSendRate, (double)perf.byteSentTotal / 1000000.0,
-	     perf.byteSentTotal ? perf.byteRetransTotal / perf.byteSentTotal * 100.0 : 0,
-	     perf.byteSentTotal ? perf.byteSndDropTotal / perf.byteSentTotal * 100.0 : 0);
+	     perf.byteSentTotal ? (double)perf.byteRetransTotal / (double)perf.byteSentTotal * 100.0 : 0,
+	     perf.byteSentTotal ? (double)perf.byteSndDropTotal / (double)perf.byteSentTotal * 100.0 : 0);
 
 	srt_epoll_release(s->eid);
 	int err = srt_close(s->fd);
